-move interfaces to types.ts -move functions to helper.ts -move variable view logic to debuggerVariables and variable view -only call scopes and variables for the first stack and scope
