+++ src/client/datascience/variablesView/variableViewProvider.ts
@@ -13,7 +13,6 @@ import {
 import { isTestExecution } from '../../common/constants';
 import { IConfigurationService, IDisposableRegistry } from '../../common/types';
 import { createDeferred, Deferred } from '../../common/utils/async';
-import { IDebuggingManager } from '../../debugger/types';
 import { Identifiers } from '../constants';
 import { IDataViewerFactory } from '../data-viewing/types';
 import { ICodeCssGenerator, IJupyterVariableDataProviderFactory, IJupyterVariables, IThemeFinder } from '../types';
@@ -57,8 +56,7 @@ export class VariableViewProvider implements IVariableViewProvider {
         private readonly jupyterVariableDataProviderFactory: IJupyterVariableDataProviderFactory,
         @inject(IDataViewerFactory) private readonly dataViewerFactory: IDataViewerFactory,
         @inject(INotebookWatcher) private readonly notebookWatcher: INotebookWatcher,
-        @inject(ICommandManager) private readonly commandManager: ICommandManager,
-        @inject(IDebuggingManager) private readonly debuggingManager: IDebuggingManager
+        @inject(ICommandManager) private readonly commandManager: ICommandManager
     ) {}
 
     public async resolveWebviewView(
@@ -81,8 +79,7 @@ export class VariableViewProvider implements IVariableViewProvider {
             this.jupyterVariableDataProviderFactory,
             this.dataViewerFactory,
             this.notebookWatcher,
-            this.commandManager,
-            this.debuggingManager
+            this.commandManager
         );
 
         // If someone is waiting for the variable view resolve that here
